job-application/backend/requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
python-multipart>=0.0.5
pydantic>=1.8.2
faiss-cpu>=1.7.4
langchain>=0.0.184
langchain-core>=0.1.0
langchain-google-genai>=0.0.3
PyPDF2
passlib[bcrypt]>=1.7.4
python-jose[cryptography]>=3.3.0
sqlalchemy>=1.4.23
asyncpg
alembic>=1.7.1
databases 
websockets 
psycopg2-binary>=2.9.1
python-dotenv>=0.19.0
requests>=2.26.0
google-generativeai>=0.3.1
stripe>=7.0.0
httpx>=0.25.0
google-cloud-talent
PyJWT 

# PDF Generation
reportlab>=4.0.0 

# Web Scraping with Playwright only
playwright==1.52.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
readability-lxml>=0.8.1 